# smart-api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  description: >-
    An project that has how the main functionality recover packings on the
    system through Beacons and Checkpoints
  title: Reciclapac API Documentation
  contact:
    name: David W. F. Santos
    email: david.santos@pe.senai.br
    url: 'http://pe.senai.br'
host: 'localhost:3000'
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
...
securityDefinitions:
  Bearer:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  '/profile/retrieve/{profile_id}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Retrieve a profile from the system
      description: ''
      operationId: profile_read
      security:
        - Bearer: []
      parameters:
        - in: path
          name: profile_id
          description: Retrieve profile related this profile id
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/retrieve/email/{email}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Retrieve a profile from the system by email
      description: ''
      operationId: profile_read_by_email
      security:
        - Bearer: []
      parameters:
        - in: path
          name: email
          description: Retrieve profile related this email
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/auth/{password}/{email}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Realyze auth on the system using password and email from profile
      description: ''
      operationId: profile_auth
      parameters:
        - in: path
          name: password
          description: Auth profile related this password
          required: true
          type: string
        - in: path
          name: email
          description: Auth profile related this email
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/recover/{password}/{email}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Realyze recover of the profile using password and email from profile
      description: ''
      operationId: profile_recover
      security:
        - Bearer: []
      parameters:
        - in: path
          name: password
          description: Recover profile related this password
          required: true
          type: string
        - in: path
          name: email
          description: Recover profile related this email
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/list/pagination/{limit}/{page}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Retrieve all profilies from the system with pagination
      description: ''
      operationId: profile_listPagination
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of profilies that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve profiles related this page
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/list/pagination/supplier/{limit}/{page}/{supplier}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Retrieve all profilies from the system with pagination by supplier
      description: ''
      operationId: profile_listPagination_supplier
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of profilies that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve profiles related this page
          required: true
          type: string
        - in: path
          name: supplier
          description: Retrieve profiles related this supplier id
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/list/pagination/logistic/{limit}/{page}/{logistic}':
    x-swagger-router-controller: profile_controller
    get:
      tags:
        - Profile
      summary: Retrieve all profilies from the system with pagination by logistic
      description: ''
      operationId: profile_listPagination_logistic
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of profiles that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve profiles related this page
          required: true
          type: string
        - in: path
          name: logistic
          description: Retrieve profiles related this logistic operator id
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /profile/create:
    x-swagger-router-controller: profile_controller
    post:
      tags:
        - Profile
      summary: Add a new profile on the system
      description: ''
      operationId: profile_create
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Sample of profile object that will be add to the system
          required: true
          schema:
            $ref: '#/definitions/Profile'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/delete/{profile_id}':
    x-swagger-router-controller: profile_controller
    delete:
      tags:
        - Profile
      summary: Remove a profile from the system
      description: ''
      operationId: profile_delete
      security:
        - Bearer: []
      parameters:
        - in: path
          name: profile_id
          description: Remove a profile related this profile id
          required: true
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/profile/update/{profile_id}':
    x-swagger-router-controller: profile_controller
    put:
      tags:
        - Profile
      summary: Update a profile from the system
      description: ''
      operationId: profile_update
      security:
        - Bearer: []
      parameters:
        - in: path
          name: profile_id
          description: Update a profile related this profile id
          required: true
          type: string
        - in: body
          name: body
          description: The profile object that will be modified
          required: true
          schema:
            $ref: '#/definitions/Profile'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /settings/retrieve:
    x-swagger-router-controller: settings_controller
    get:
      tags:
        - Settings
      summary: Retrieve settings from the system
      description: ''
      operationId: settings_retrieve
      security:
        - Bearer: []
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /settings/update:
    x-swagger-router-controller: settings_controller
    put:
      tags:
        - Settings
      summary: update settings from the system
      description: ''
      operationId: settings_update
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: The setting object that will be modified
          required: true
          schema:
            $ref: '#/definitions/Setting'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  ErrorResponse:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
  Error400:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
  Error500:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
...
  Profile:
    type: object
    properties:
      profile:
        type: string
      password:
        type: string
      email:
        type: string
      city:
        type: string
      street:
        type: string
      user:
        type: string
      telephone:
        type: string
      cellphone:
        type: string
      cep:
        type: string
      neighborhood:
        type: string
      uf:
        type: string
...
  Setting:
    type: object
    properties:
      battery_level:
        type: number
      clean:
        type: number
      range_radius:
        type: number
        format: double
      register_gc16:
        type: object
        properties:
          enable:
            type: boolean
          id:
            type: string
  SuccessResponse:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
    required:
      - jsonapi
...